changed test for double to numeric

+++ R/r_gsl_odeiv2.R
@@ -146,7 +146,7 @@ r_gsl_odeiv2_outer <- function(name,experiments,p,abs.tol=1e-6,rel.tol=1e-5,init
 	stopifnot(file.exists(so))
 	stopifnot(is.matrix(p))
 	stopifnot(any(c('outputTimes','time') %in% names(experiments[[1]])))
-	stopifnot(is.double(experiments[[1]]$outputTimes) || is.double(experiments[[1]]$time))
+	stopifnot(is.numeric(experiments[[1]]$outputTimes) || is.numeric(experiments[[1]]$time))
 	y <- .Call(odeiv_outer_e,name,experiments,p,abs.tol,rel.tol,initial.step.size)
 	return(y)
 }
